pyproject.toml
[tool.poetry]
name = "moot-app"
version = "0.1.0"
description = "A simple app for taking moot bookings"
authors = []

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.2"
python-dotenv = "^0.14.0"
WTForms = "^2.3.3"
wtforms = {extras = ["email"], version = "^2.3.3"}
email_validator = "^1.1.2"
gunicorn = "^20.0.4"
Flask-WTF = "^0.14.3"
requests = "^2.25.1"
psycopg2-binary = "^2.8.6"
Flask-Mobility = "^1.1.0"
sendgrid = "^6.7.0"
Flask-Table = "^0.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
